test(notification-handler): add the notification handler test mock to throw domain exception on call validate and add notification handler test unit in domain layer to test units
